Merge `parser` and `grammar` packages into single `omgidl-parser` (#26)
Since these packages are so tightly coupled and it doesn't make sense to
serve the `grammar` out of it's own package, I've consolidated these
into a single parser package. I've limited the internal AST types and
parsing to not be exported as part of the module because nothing we have
needs direct access to creating the AST yet. I wanted to keep as much
private as possible for now.

Updated the readmes, package.jsons, and all references to `-grammar` in
accordance with these changes.


https://linear.app/foxglove/issue/FG-4239/merge-grammar-and-parser-packages