Deserialization support for Unions (#64)
The message reader can now read union types. I broke out the
functionality from the main `readComplexType` function into 4 other
methods for specifically reading fields, unions, structs, and aggregate
types.

Also when an enum is the switchType of a union, that allows the cases to
refer only to their member enumerators directly without having to
namespace them to the type, so I had to update the logic in UnionIDLNode
to allow for that.