Fix not reading dHeaders for aggregated types within sequences or arrays (#74)
Previously I thought that `DHEADER`s were only read in for top level
types, but it turns out they should be read in for all aggregated types
(unions and structs) when their size hasn't been read in already by an
`EMHEADER`. This can happen when these aggregated types are nested
within an array. I've changes `isTopLevel` option which previously
controlled the reading in of the DHEADER to be `knownTypeSize` to
indicate whether the size of the aggregated type is already known. If it
is known, then we don't read in the DHEADER.

Applies to PL_CDR2 encapsulation types.