Improve omgidl deserialization performance (#123)
### Public-Facing Changes

Improve omgidl deserialization performance

### Description
This PR changes the `MessageReader` to build a tree structure of
deserialization info objects when constructed. When building this tree,
the concrete deserializers as well as other relevant type information
are looked up which avoids that this has to be done during the actual
deserialization. Further changes include the removal of some temporary
closures which reduces some GC time.

The outcome of these changes is a performance improvement of up to 30%
(depends on message type)


![image](https://github.com/foxglove/omgidl/assets/9250155/84e77541-f673-43fd-8c63-4fa0d4bd974f)
In the image above, **left** is this PR, **right** is the currently
released version

Resolves FG-6300