Fix edge cases of optional parameters (#217)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- Fix for reading optional fields in final structs, we now expect their
emHeaders
- Fix for reading absent optional fields at the end of a message that
would cause a RangeError

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description


Fixes two issues in our implementation of optional fields. 

The first issue (FG-8644) was us not expecting optional parameters of
final structs to have emHeaders in XCDR1.
This is specified in the spec here: 
<img width="595" alt="image"
src="https://github.com/user-attachments/assets/3213fa6f-9c0e-4019-afe1-3e0c7f6fe9d4">
<img width="629" alt="image"
src="https://github.com/user-attachments/assets/76b09dd7-1637-4515-99e9-b0f1ab675d57">

This was fixed by adding `&& field.isOptional`.

The second issue (FG-8627) fixed occurred in PL_CDR2 when trying to read
the absent, optional, final field of a message, which was causing a
RangeError because emHeaders are not present for absent optional fields
in PL_CDR2 since they are able to check id's of the emHeaders. This was
fixed by catching the `RangeError` and returning `undefined` as we do
when the emHeader is not the id we expect which is the other case when
optional field is not present.

I also wrapped the `readMemberFieldValue` in a `try/catch` to be able to
give better error messages at which field failed and where in the
serialized message it encountered the error.