Allow for Mutable and Appendable types to really work across schemas.  (#321)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

 - Mutable structs now support out of order members
- Appendable structs support absence of members in schema (old schema
new message still allows for reading, same with vice versa)

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

Background on XCDR Appendable vs Mutable extensibility types. 

When an aggregated type (`struct` or `union` ) is `mutable` it means
that the struct may be changed across versions by adding/removing
members or moving the members around within the schema. Given that these
rules are followed and the type was always mutable, this should mean
that new schemas are able to read old messages and vice versa without
failing (not all data will be available though). `id`'s should not be
changed across mutable types however.

Appendable aggregated types are only expected to be added to over time.
Meaning that the fields are always in the same order with greater /
fewer fields across time.

The final extensibility type is `Final` which should not change at all
over time.

The encoding of parts of messages with these extensibility types can
change depending on whether it is XCDR2 or not.

References to all aggregated types encoding pattern is below:

## Changes to how we handle mutable structs

<img width="605" height="208" alt="image"
src="https://github.com/user-attachments/assets/2506b8e4-f2d6-419e-a67a-a06dbd4f1c46"
/>

> <em>Mutable Struct XCDR1</em>

<img width="661" height="375" alt="image"
src="https://github.com/user-attachments/assets/2af0bdd2-1cf9-411b-ab79-83865965b6cc"
/>

> <em>Mutable Struct XCDR2</em>

Previously we were scanning through the members of each struct and
expecting them to match the ids in the emHeaders as we found them, and
if the emHeaders weren't present we would assume that field was missing
and save that last read, unused emHeader for the next time we were meant
to read it. This approach doesn't work for out of order struct members
because we were expecting the emHeader order to match the struct field
order or otherwise be absent.

A few major changes were made to allow for this.

The first major change is that we don't loop over the struct by the
order of fields. We now read the emHeader from the message to determine
what field we should be expecting, and use the `id` in the emHeader to
find the field in the message. To facilitate this I added to fields to
the `deserializationInfo` for structs: `fieldsInOrder` and `fieldsById`.

Initially `fieldsById` was only reading from explicitly defined `ids`
for each field, when in reality, if the field doesn't have the `id`
annotation it should receive an implied `id` based on its order in the
message with some exceptions around explicitly defined ids. I added the
setting of these implied ids after the setting of the explicit ids in
the `fieldsById` map and did it according to spec: sequentially starting
at 0, skipping preset explicit `ids`.

<img width="652" height="914" alt="image"
src="https://github.com/user-attachments/assets/d06dedbf-03cb-4288-9122-8ceca8b73af9"
/>

> _reference for implied ids for mutable structs_


Now that we can read members out of order we need to know when to stop
reading emHeaders so that we don't read outside of the struct. To do
this we now take into consideration the `Dheader` present on all XCDR2
Mutable structs, and the SENTINEL_PID which signifies the end of all
XCDR1 mutable structs. The DHeader denotes the size of the struct, so we
know when we've reached that offset from the start of the struct that
the struct is over. Then for the sentinelHeader, we do have the ability
to read that alongside the emheader, but I chose to have it check to see
if the next reading is a sentinel header and reset the reader back to
the original offset if it doesn't encounter it. These two cases give us
our exit case from reading the mutable struct.

After reading what we can of the mutable struct from the message, we now
have to fill in any missing optional or undefined fields we were
expecting. Now we track what members were written to the message, so
that we can iterate through the rest of the fields that weren't written
to the message and ensure that they are assigned to their proper default
values.

## Changes to Appendable and Final Structs


<img width="638" height="349" alt="image"
src="https://github.com/user-attachments/assets/2f0ff3c6-5050-405f-a1d4-f850f1a034ef"
/>

> <em>Appendable modifies Final Structs and Unions in this way</em>

<img width="665" height="199" alt="image"
src="https://github.com/user-attachments/assets/53127e66-1f50-4b5d-a3b8-aebe7816eb6e"
/>

> <em>Final Struct XCDR 1&2</em>

Two main issues with how we previously handled final/appendable structs:
1. We didn't stop reading when we reached the Dheader (in the case of a
new schema reading an old messages with appendable extensibility type)
 2. We weren't handling final optional members properly.


<img width="596" height="504" alt="image"
src="https://github.com/user-attachments/assets/bfe97b77-a598-473b-a5ca-a38376fb1188"
/>

> <em>While they have the same outer structure for v1 and v2, the
encoding of the optional members differ between XCDR1 & XCDR2:</em>

To fix this we are properly checking the is_present flag for XCDR2
structs instead of just checking EMHEADER which should only be used for
XCDR1 OPT_FMEMBERS.

Other minor issues:
- we weren't setting defaults for missing fields in appendable type
structs
- we now detect end of message and fill in any missing fields from the
schema
- we weren't anticipating for messages that could contain more fields
than what the schema had
- in this case we would have run out of fields in the struct but there
is still data left in the struct, so we skip to the end fo the struct
using the dheader. (Should also probably support this for XCDR1 🤔 )

## Changes to handling Final/Appendable Unions

<img width="660" height="238" alt="image"
src="https://github.com/user-attachments/assets/2b1a2657-879d-42e5-8a58-41079cc5bc66"
/>

> _ Final/Appendable union serialization_

I think the only thing that's really changed here is the using the
is_present flag on OPT_FMEMBERS.


## Overall changes:
- Can properly detect that structs and unions have ended by using their
Dheaders and sentinel flags
 - more controlled usage of sentinel flags
 - Allow for the reading of mutable structs out of order
 - proper assignment of implied `id`s for members of mutable structs
 - support backward compatibility of appendable structs
- support proper encoding of OPT_FMEMBERS in XCDR2 using is_present
flag.

Test changes:
- Some tests used dummy dheaders -- these have been updated to use
actual values that can be used so that it doesn't terminate the struct
early/late

Off topic changes:
 - fixed jest vscode integration
- consistent naming of `usesDelimiterHeaders` from
`readDelimiterHeaders` (same for EmHeaders)

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->
Fixes:
https://linear.app/foxglove/issue/FG-12378/appending-fields-to-appendablemutable-structs-breaks-xcdr2-decoding
Fixes: #313